Move game part to Game.vue component
