+++ src/components/Ranking.vue
@@ -1,6 +1,6 @@
 <template>
   <div class="container">
-    <h1>hello</h1>
+    <h1>hello ranking</h1>
   </div>
 </template>
